add: clip screen logic

+++ features/home/ui/src/main/java/ru/popkov/russeliterature/features/home/ui/HomeViewAction.kt
@@ -1,7 +1,7 @@
 package ru.popkov.russeliterature.features.home.ui
 
 sealed interface HomeViewAction {
-    data class OnCardClick(val cardId: Int) : HomeViewAction
+    data class OnCardClick(val cardId: Long) : HomeViewAction
     data class OnSectionClick(val sectionId: Int) : HomeViewAction
     data class OnAuthorFaveClick(val userId: Long, val authorId: Long, val isFave: Boolean) : HomeViewAction
     data class OnArticleFaveClick(val userId: Long, val articleId: Long, val isFave: Boolean) : HomeViewAction
